requirements.txt
Flask==0.12.1
RPi.GPIO==0.6.5
